Created parameters module
Moved functionality for building calculation parameters from `cost` into
new `parameters` module.
Moved building parameters outside building cost matrix function.
